requirements.txt
requests>=2.31.0
requests-oauthlib>=1.3.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
Pillow>=10.2.0
python-slugify>=8.0.1
python-dateutil>=2.8.2
tqdm>=4.66.1
coloredlogs>=15.0.1
